Integration of rules_parser in launcher
