+++ qos.py
@@ -9,18 +9,10 @@ import sys
 import logging
 
 try:
-    from config import INTERFACES
-except ImportError:
-    print("No existing configuration. Please copy config.py.default as "
-          "config.py and optionaly configure it for your setup.")
-    exit(1)
-
-try:
     from config import DEBUG
 except ImportError:
     DEBUG = False
 
-import rules
 import tools
 
 
@@ -36,29 +28,19 @@ def run_as_root():
         exit()
 
 
-def get_ifnames(interfaces_lst=INTERFACES):
-    if_names = set()
-    for interface in interfaces_lst.values():
-        if "name" in interface.keys():
-            if_names.add(interface["name"])
-        else:
-            if_names.update(get_ifnames(interfaces_lst=interface))
-    return if_names
-
-
 def apply_qos():
     run_as_root()
     # Clean old rules
     reset_qos()
     # Setting new rules
-    logging.info("Setting new rules")
+    print("Setting new rules")
 
-    rules.apply_qos()
+    setup_qos()
 
 
 def reset_qos():
     run_as_root()
-    logging.info("Removing tc rules")
+    print("Removing tc rules")
     ifnames = get_ifnames()
     tools.qdisc_del(ifnames, "htb", stderr=subprocess.DEVNULL)
     return
@@ -76,7 +58,7 @@ def set_debug(level):
     if level or DEBUG:
         log_level = logging.DEBUG
     else:
-        log_level = logging.INFO
+        log_level = logging.WARNING
 
     logging.basicConfig(
         stream=sys.stderr,
@@ -89,18 +71,32 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser(
         description="Script to set, show or delete QoS rules with TC"
     )
-    sp = parser.add_subparsers()
-    sp_start = sp.add_parser("start", help="set QoS rules")
-    sp_stop = sp.add_parser("stop", help="Remove all QoS rules")
-    sp_show = sp.add_parser("show", help="Show QoS rules")
+
+    # Start/Stop/Show command
+    sp_action = parser.add_subparsers()
+    sp_start = sp_action.add_parser("start", help="set QoS rules")
+    sp_stop = sp_action.add_parser("stop", help="Remove all QoS rules")
+    sp_show = sp_action.add_parser("show", help="Show QoS rules")
 
     # Set function to call for each options
     sp_start.set_defaults(func=apply_qos)
     sp_stop.set_defaults(func=reset_qos)
     sp_show.set_defaults(func=show_qos)
 
+    # Debug option
     parser.add_argument('-d', '--debug', help="Set the debug level",
                         dest="debug", action="store_true")
+
+    # Different ways to create QoS
+    parser_group = parser.add_mutually_exclusive_group()
+
+    # Use class rules
+    parser_group.add_argument('-c', '--class',
+                              help="Use class rules (default)",
+                              dest="class", action="store_true")
+    # Use tree rules
+    parser_group.add_argument('-t', '--tree', help="Use tree rules",
+                              dest="tree", action="store_true")
 
     # If no argument provided show help
     if len(sys.argv) == 1:
@@ -112,6 +108,11 @@ if __name__ == '__main__':
 
     # Set debug mode
     set_debug(args.debug)
+
+    if args.tree:
+        from rules_parser import setup_qos, get_ifnames
+    else:
+        from class_parser import setup_qos, get_ifnames
 
     # Execute correct function
     args.func()
